Deprecate the ``at_bbb`` and ``dx_bbb`` modules. They are merged with the ``at`` respectively ``dx`` modules.
